Update Schema and Script
Updated the schema then updated the script to fully reflect the schema. Also had to make some changes to the team_stats function in the library so that the names of the columns match up to SQL best practices using snake case and descriptive names.

The script should be ready to run for all of NCAA now.